Refactor: extract chunk type functionality into separate file
Move chunk type determination logic from cmp_icu.c into its own file
